fix: renamed journey url to journey link
