+++ apps/journeys-admin/src/components/JourneyView/Properties/JourneyLink/EmbedJourneyDialog/EmbedCardPreview/EmbedCardPreview.tsx
