--- apps/journeys-admin/src/components/JourneyView/Properties/JourneyURL/index.ts
@@ -1,1 +1,0 @@
-export { JourneyURL } from './JourneyURL'
